+++ python/GafferImageUI/ImageViewUI.py
@@ -81,8 +81,6 @@ Gaffer.Metadata.registerNode(
 	"toolbarLayout:customWidget:BottomRightSpacer:section", "Bottom",
 	"toolbarLayout:customWidget:BottomRightSpacer:index", -1,
 
-	"toolbarLayout:activator:gpuAvailable", lambda node : isinstance( GafferImageUI.ImageView.createDisplayTransform( node["displayTransform"].getValue() ), GafferImage.OpenColorIOTransform ),
-
 	plugs = {
 
 		"view" : [
@@ -238,22 +236,6 @@ Gaffer.Metadata.registerNode(
 			"presetNames", lambda plug : IECore.StringVectorData( GafferImageUI.ImageView.registeredDisplayTransforms() ),
 			"presetValues", lambda plug : IECore.StringVectorData( GafferImageUI.ImageView.registeredDisplayTransforms() ),
 
-		],
-
-		"lutGPU" : [
-			"description",
-			"""
-			Controls whether to use the fast GPU path for applying exposure, gamma, and displayTransform.
-			""",
-
-			"toolbarLayout:activator", "gpuAvailable",
-
-			# Turning off GPU mode means we can't properly support efficient wipes.  Since we no longer
-			# have feature parity, we're deprecating CPU mode, and expecting everyone to use the GPU path.
-			# With new OCIO, the GPU path now gives high quality on any even vaguely recent GPU.
-			# The only use case I can think of for the CPU path is checking that the GPU path is working correctly.
-			# If you need to do that, set this metadata to True, and then create a new Viewer to force a refresh.
-			"toolbarLayout:visibilityActivator", False,
 		],
 
 		"colorInspector" : [